refactor: reflect code review
