+++ src/main/java/step3/domain/CarNumber.java
@@ -1,9 +1,5 @@
 package step3.domain;
 
-import java.util.List;
-import java.util.stream.Collectors;
-import java.util.stream.IntStream;
-
 import step3.utils.ValidationUtils;
 
 public class CarNumber {
@@ -14,10 +10,7 @@ public class CarNumber {
 		this.number = number;
 	}
 
-	public List<Car> createCars() {
-		return IntStream.range(0, number)
-			.boxed()
-			.map(integer -> new Car())
-			.collect(Collectors.toList());
+	public int getNumber() {
+		return number;
 	}
 }
