+++ src/main/java/step3/domain/RacingCarGame.java
@@ -1,17 +1,19 @@
 package step3.domain;
 
+import step3.utils.RandomNumberGenerator;
+
 public class RacingCarGame {
 	private final Cars cars;
 	private final TryNumber tryNumber;
 
 	public RacingCarGame(CarNumber carNumber, TryNumber tryNumber) {
-		this.cars = new Cars(carNumber.createCars());
+		this.cars = new Cars(carNumber);
 		this.tryNumber = tryNumber;
 	}
 
 	public void play() {
 		tryNumber.play();
-		cars.goOrStop();
+		cars.goOrStop(RandomNumberGenerator.getInstance());
 	}
 
 	public Cars getCars() {
