+++ src/test/java/step3/domain/CarsTest.java
@@ -11,7 +11,7 @@ public class CarsTest {
 	@Test
 	@DisplayName("Cars의 Car 리스트는 수정할 수 없다.")
 	void unmodifiableListTest() {
-		List<Car> cars = new Cars(new CarNumber(3).createCars()).getCars();
+		List<Car> cars = new Cars(new CarNumber(3)).getCars();
 
 		assertThatExceptionOfType(UnsupportedOperationException.class)
 			.isThrownBy(() -> cars.add(new Car()));
@@ -20,16 +20,18 @@ public class CarsTest {
 	@Test
 	@DisplayName("항상 전진하는 Cars를 생성하고 모든 Car가 전진하는지 확인한다.")
 	void goOrStopTest() {
-		Cars cars = new Cars(new CarNumber(3).createCars()) {
-			@Override
-			public int getNumber() {
-				return 4;
-			}
-		};
+		Cars cars = new Cars(new CarNumber(3));
 
-		cars.goOrStop();
+		cars.goOrStop(() -> 4);
 
 		assertThat(cars.getCars())
 			.containsExactly(new Car(1), new Car(1), new Car(1));
+	}
+
+	@Test
+	@DisplayName("입력된 carNumber만큰 Car 리스트를 생성한다.")
+	void createCarsTest() {
+		int number = 3;
+		assertThat(Cars.createCars(new CarNumber(number))).hasSize(number);
 	}
 }
